Making the FMOD logo look nicer
I want to position it so each game has it at roughly the same coords.  I think it looks better under the final menu item as when you ESC mid-game, it looks a bit odd in the bottom-left.

Though this is just temporary until I can be bothered to code a splash screen or 'about' window in the launcher.