Cleanups and fixes for macOS SDL2
The SDL2 backend on macOS was borked.  The timer did not correctly quit correctly and so it always exited without the ending sounds.  I'm using the i_timer for cocoa while compiling with SDL2 so it fixes it.  SDL2 with Cocoa is the best solution.  Clearly whoever wrote the SDL2 timer didn't test it on macOS.