Removing this feature on a Mac
On Modern Macs, using SDL2 this won't really work as it will conflict with other bindings on macOS.  It will work in Cocoa but then the Cocoa backend has an OpenGL framebuffer, so really for future proofing, SDL2 is the way to go.

Who really wants to switch mid-game anyway?  It will size the window to the same resolution you chose, so you have to go into the menu to fix that anyway!