Dimming all menus after ESC
Keeps the titlepic etc..  but the moment you hit a key to bring up the menu, it'll overdim 10 points to pitch black.  This allows you to actually read the text on the various menus instead of it contrasting badly..  even I have problems reading that sometimes.  Whoever thought this was readable must have been completely out of their mind!